Compiler warnings Remove print_html_auras_buffs() due to r10732
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@10740 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_player.cpp
@@ -6832,7 +6832,7 @@ targetdata_t* targetdata_t::get( player_t* source, player_t* target )
   if( id < 0 )
     source -> targetdata_id = id = source -> sim -> num_targetdata_ids++;
 
-  if( id >= target -> targetdata.size() )
+  if( id >= ( int ) target -> targetdata.size() )
     target -> targetdata.resize( id + 1 );
 
   targetdata_t* p = target->targetdata[id];
